grafcet-agents/requirements.txt
fastapi
uvicorn
google-generativeai
google-genai
pydantic
python-dotenv
websockets
aiohttp
aiofiles
pypdf

